Ag 5/requirements.txt
Flask
mysql-connector-python
python-dotenv
Flask-Mail
gunicorn
